requirements.txt
openai
whisper
transformers
streamlit
gtts
speechrecognition
pyaudio
vosk
llama-cpp-python